address review: rename to converter.py, inherit Converter, pass converters list directly
- rename process.py -> converter.py
- ModelFreePtqConverter(Converter) explicit inheritance
- validate: remove redundant .to("meta"), tensors are already meta per protocol
- update_config: rename incoming -> config to match base class signature
- _process_shard/_validate_shard: accept converters list[Converter] directly
  instead of rebuilding from config+converter each shard; device now passed
  through in _validate_shard instead of hardcoded
